Assistant checkpoint: Added Air Force standards and K-factors
Assistant generated file changes:
- standards_db.py: Add Air Force standards
- analysis_engine.py: Add Air Force specific K-factors

---

User prompt:

There is also a supplement to the DoD STD for Air Force. Make sure that is in there and all the values are loaded. Small things can be different and I want to ensure when they select the databse, Air Force can have it's perfect enviornment. Let's go over ALL things done so far from the 1st thing of this project.