Assistant checkpoint: Added NATO standards and unit conversion support
Assistant generated file changes:
- standards_db.py: Add NATO standards, Add NATO references
- analysis_engine.py: Add NATO k-factors and unit conversion, Add NATO factors, Add unit conversion to calculations, Add unit conversion logic

---

User prompt:

Also, we need to make an addition and include our NATO friends. So we will also be giving the user the option to do qd analysis for NATO. Also, the DoE labs use lbs and g and kg for their calculations. Don't forget the small things like this.