Assistant checkpoint: Remove extra drawing controls after clicking buttons
Assistant generated file changes:
- static/js/ui-controls.js: Remove extra drawing controls when clicking on polygon, Remove extra drawing controls when clicking on rectangle, Remove extra drawing controls when clicking on marker, Add initialization to prevent toolbars

---

User prompt:

I really don't know how else to tell you to fix this. Nothing you are doing is working to fix this issue. When I click on "Draw Polygon" it opens a smaller icon on the map called "Draw a Polygon". I don't want that to happen. I only want to click on "Draw Polygon" and have the ability to draw. Understand?